move brepdata string and brep creation utility out to own file and import from transformer tests
